Changed Restaurant to have a menuID and moved a copy to newbackend
