+++ src/main/java/cs/dal/csci3130/undined/newbackend/Restaurant.java
@@ -1,4 +1,4 @@
-package cs.dal.csci3130.undined.backend;
+package cs.dal.csci3130.undined.newbackend;
 
 import org.apache.commons.beanutils.BeanUtils;
